feat: add server-side tRPC prefetching with HydrationBoundary
