+++ package.json
@@ -47,6 +47,7 @@
     "react-markdown": "^10.1.0",
     "recharts": "^3.7.0",
     "remove-markdown": "^0.6.3",
+    "server-only": "^0.0.1",
     "shiki": "^3.22.0",
     "slug": "^11.0.1",
     "stale-lru-cache": "^5.1.1",
