+++ src/app/[place]/armor/page.tsx
@@ -1,17 +1,25 @@
-'use client';
-
+import { dehydrate, HydrationBoundary } from '@tanstack/react-query';
 import React from 'react';
 
-import ArmorVisualizer from '@/components/features/tools/armor/armorVisualizer';
-import Layout from '@/components/layout/layout';
-import PageMeta from '@/components/layout/pageMeta';
+import { prefetchPlace } from '@/trpc/server';
 
-export default function ArmorPage() {
+import ArmorClient from './_client';
+
+export default async function ArmorPage({
+  params,
+}: {
+  params: Promise<{ place: string }>;
+}) {
+  const { place: initials } = await params;
+  const { helpers, place } = await prefetchPlace(initials);
+
+  if (place) {
+    await helpers.vehicles.list.prefetch({ placeId: place.placeId });
+  }
+
   return (
-    <PageMeta title="Armor visualizer">
-      <Layout hidePlaceDropdown noPadding overwriteTabLabel="Armor visualizer">
-        <ArmorVisualizer />
-      </Layout>
-    </PageMeta>
+    <HydrationBoundary state={dehydrate(helpers.queryClient)}>
+      <ArmorClient />
+    </HydrationBoundary>
   );
 }
