+++ src/app/[place]/teams/[team]/page.tsx
@@ -1,72 +1,30 @@
-'use client';
-
-import { Flex, Stack } from '@chakra-ui/react';
-import { useParams, useRouter } from 'next/navigation';
+import { dehydrate, HydrationBoundary } from '@tanstack/react-query';
 import React from 'react';
-import { GrDocumentMissing } from 'react-icons/gr';
 import slugify from 'slug';
 
-import TeamHeader from '@/components/features/teams/header';
-import TeamLoadouts from '@/components/features/teams/loadouts';
-import Layout from '@/components/layout/layout';
-import PageMeta from '@/components/layout/pageMeta';
-import { EmptyState } from '@/components/ui/empty-state';
-import { usePlace } from '@/hooks/usePlace';
-import { trpc } from '@/utils/trpc';
+import { prefetchPlace } from '@/trpc/server';
 
-export default function PlaceTeam() {
-  const router = useRouter();
-  const params = useParams();
-  const teamQuery = params.team as string;
-  const teamSlug = slugify(teamQuery);
-  const place = usePlace()!;
+import PlaceTeamClient from './_client';
 
-  const [team] = trpc.teams.bySlug.useSuspenseQuery({
-    placeId: place.placeId,
-    slug: teamSlug,
-  });
+export default async function PlaceTeamPage({
+  params,
+}: {
+  params: Promise<{ place: string; team: string }>;
+}) {
+  const { place: initials, team } = await params;
+  const { helpers, place } = await prefetchPlace(initials);
 
-  React.useEffect(() => {
-    if (!team) return;
-
-    if (teamQuery !== teamSlug) {
-      const newPath = `/${place.initials}/teams/${teamSlug}`;
-      router.replace(newPath);
-    }
-  }, [router, teamQuery, team, teamSlug, place.initials]);
-
-  const title = team ? team.name : 'Team not found';
-  const description = team
-    ? `${team.name} team statistics and vehicle compositions for ${place.placeName}`
-    : undefined;
+  if (place) {
+    const teamSlug = slugify(team);
+    await helpers.teams.bySlug.prefetch({
+      placeId: place.placeId,
+      slug: teamSlug,
+    });
+  }
 
   return (
-    <PageMeta title={title} description={description}>
-      <Layout noPadding>
-        {team ? (
-          <Flex
-            justifyContent="center"
-            marginBottom={{ base: 4, md: 0 }}
-            padding={{ base: 0, md: 2, lg: 4 }}
-          >
-            <Stack
-              aria-labelledby="team-page-title"
-              as="article"
-              data-md-target
-              gap={4}
-              maxWidth="5xl"
-              width="100%"
-            >
-              <TeamHeader initials={place.initials} name={team.name} />
-              <TeamLoadouts initials={place.initials} team={team} />
-            </Stack>
-          </Flex>
-        ) : (
-          <Flex alignItems="center" height="100%">
-            <EmptyState icon={<GrDocumentMissing />} title="Team not found" />
-          </Flex>
-        )}
-      </Layout>
-    </PageMeta>
+    <HydrationBoundary state={dehydrate(helpers.queryClient)}>
+      <PlaceTeamClient />
+    </HydrationBoundary>
   );
 }
