+++ src/app/[place]/vehicles/page.tsx
@@ -1,25 +1,25 @@
-'use client';
-
-import { Flex } from '@chakra-ui/react';
+import { dehydrate, HydrationBoundary } from '@tanstack/react-query';
 import React from 'react';
-import { GrDocumentMissing } from 'react-icons/gr';
 
-import VehiclesSearchSidebar from '@/components/features/vehicles/searchSidebar';
-import Layout from '@/components/layout/layout';
-import SearchLayout from '@/components/layout/searchLayout/layout';
-import { EmptyState } from '@/components/ui/empty-state';
+import { prefetchPlace } from '@/trpc/server';
 
-export default function PlaceVehicles() {
+import PlaceVehiclesClient from './_client';
+
+export default async function PlaceVehiclesPage({
+  params,
+}: {
+  params: Promise<{ place: string }>;
+}) {
+  const { place: initials } = await params;
+  const { helpers, place } = await prefetchPlace(initials);
+
+  if (place) {
+    await helpers.vehicles.list.prefetch({ placeId: place.placeId });
+  }
+
   return (
-    <Layout noPadding>
-      <SearchLayout sidebar={<VehiclesSearchSidebar />}>
-        <Flex alignItems="center" height="100%">
-          <EmptyState
-            icon={<GrDocumentMissing />}
-            title="No vehicle selected"
-          />
-        </Flex>
-      </SearchLayout>
-    </Layout>
+    <HydrationBoundary state={dehydrate(helpers.queryClient)}>
+      <PlaceVehiclesClient />
+    </HydrationBoundary>
   );
 }
