+++ src/app/[place]/winrate/page.tsx
@@ -1,95 +1,25 @@
-'use client';
-
-import { Code, Flex, Grid, Span, Stack } from '@chakra-ui/react';
-import { useQueryState } from 'nuqs';
+import { dehydrate, HydrationBoundary } from '@tanstack/react-query';
 import React from 'react';
-import slug from 'slug';
 
-import SimpleSelect from '@/components/common/simpleSelect';
-import WinrateChartRoot from '@/components/features/winrate/chart/root';
-import Layout from '@/components/layout/layout';
-import { Alert } from '@/components/ui/alert';
-import { usePlace } from '@/hooks/usePlace';
-import { slugifyArray } from '@/utils/slugifyArray';
-import { trpc } from '@/utils/trpc';
+import { prefetchPlace } from '@/trpc/server';
 
-const EMPTY_ARRAY: string[] = [];
+import PlaceWinrateClient from './_client';
 
-export default function PlaceWinrate() {
-  const place = usePlace()!;
-  const [loadout, setLoadout] = useQueryState('loadout');
-  const [map, setMap] = useQueryState('map');
+export default async function PlaceWinratePage({
+  params,
+}: {
+  params: Promise<{ place: string }>;
+}) {
+  const { place: initials } = await params;
+  const { helpers, place } = await prefetchPlace(initials);
 
-  const [winrateMetadata] = trpc.winrate.metadata.useSuspenseQuery({
-    placeId: place.placeId,
-  });
-
-  const loadoutOptions = winrateMetadata?.loadout ?? EMPTY_ARRAY;
-  const mapOptions = winrateMetadata?.map ?? EMPTY_ARRAY;
-
-  const [loadoutSlugs, mapSlugs] = React.useMemo(() => {
-    return [slugifyArray(loadoutOptions), slugifyArray(mapOptions)];
-  }, [loadoutOptions, mapOptions]);
-
-  const firstLoadout = loadoutOptions[0] ?? null;
-  const firstLoadoutSlug = React.useMemo(
-    () => (firstLoadout ? slug(firstLoadout) : null),
-    [firstLoadout],
-  );
-
-  React.useEffect(() => {
-    if (!firstLoadoutSlug) return;
-
-    if (!loadout || !loadoutSlugs[loadout]) setLoadout(firstLoadoutSlug);
-  }, [firstLoadoutSlug, loadout, loadoutSlugs, setLoadout]);
-
-  const actualLoadout =
-    loadout && loadoutSlugs[loadout] ? loadoutSlugs[loadout] : null;
-  const actualMap = map && mapSlugs[map];
-
-  const resolvedLoadout = actualLoadout ?? firstLoadout ?? null;
+  if (place) {
+    await helpers.winrate.metadata.prefetch({ placeId: place.placeId });
+  }
 
   return (
-    <Layout>
-      <Flex justifyContent="center">
-        <Stack as="main" gap={4} maxWidth="2xl" width="100%">
-          <Grid gap={2} templateColumns="repeat(2, 1fr)">
-            <SimpleSelect
-              items={loadoutOptions}
-              label="Loadout"
-              allowEmpty={false}
-              value={resolvedLoadout}
-              onValueChange={(value) => setLoadout(value ? slug(value) : null)}
-            />
-            <SimpleSelect
-              items={mapOptions}
-              label="Map"
-              noValueLabel="All"
-              value={actualMap}
-              onValueChange={(value) => setMap(value ? slug(value) : null)}
-            />
-          </Grid>
-
-          <WinrateChartRoot loadout={resolvedLoadout} map={actualMap} />
-
-          <Alert
-            background="bg.subtle"
-            borderStartColor="blue.600"
-            borderStartWidth={4}
-            colorPalette="gray"
-            startElement
-            title="Calculation"
-            marginTop={4}
-          >
-            Daily team winrate (not cumulative). For each day, the winner is the
-            team with the highest final score (ties pick the first).{' '}
-            <Span whiteSpace="nowrap">
-              Winrate = <Code>wins / games x 100</Code>
-            </Span>{' '}
-            for that day.
-          </Alert>
-        </Stack>
-      </Flex>
-    </Layout>
+    <HydrationBoundary state={dehydrate(helpers.queryClient)}>
+      <PlaceWinrateClient />
+    </HydrationBoundary>
   );
 }
